Conflicts:
	examples/src/Main.cpp
	framework/src/minko/render/IndexBuffer.cpp
	framework/src/minko/render/VertexBuffer.cpp

Former-commit-id: 9f6498659855aea17551e486a8c1c580c6a7cfa1 [formerly dffee97e18ecd9cca4dc25bed8f1bfdf479cb142] [formerly c8e98add4703dbd9f1a09cf727ec5d0ce85e8913]
Former-commit-id: df810d04e2ffb1628679950ef09bb46b385d12f1 [formerly 9d9b91597e7532f025183ffc7d3180ca55c4d900] [formerly f738b524f74339c5f8e3fce273ceb1099a59c267]
Former-commit-id: e4a9e778c49b4782b3dd98687866c59489b358de
Former-commit-id: 1c74362e607a34d938f0b1cedb1c1ace73e7d0bb

+++ framework/src/minko/render/Texture.cpp
@@ -25,10 +25,12 @@ using namespace minko::render;
 
 Texture::Texture(std::shared_ptr<render::AbstractContext>	context,
 				 const unsigned int							width,
-				 const unsigned int							height) :
+				 const unsigned int							height,
+                 bool                                       optimizeForRenderToTexture) :
 	AbstractResource(context),
 	_width(width),
-	_height(height)
+	_height(height),
+    _optimizeForRenderToTexture(optimizeForRenderToTexture)
 {
 }
 
@@ -62,13 +64,19 @@ Texture::data(unsigned char* data, DataFormat format)
 void
 Texture::upload()
 {
-	_id = _context->createTexture(_width, _height, false);
-	_context->uploadTextureData(_id, _width, _height, 0, &_data[0]);
+    if (_id == -1)
+    	_id = _context->createTexture(_width, _height, false, _optimizeForRenderToTexture);
+	
+    if (_data.size())
+        _context->uploadTextureData(_id, _width, _height, 0, &_data[0]);
 }
 
 void
 Texture::dispose()
 {
-	_context->deleteTexture(_id);
-	_id = -1;
+    if (_id != -1)
+    {
+	    _context->deleteTexture(_id);
+	    _id = -1;
+    }
 }
